requirements.txt
streamlit==1.28.0
numpy==1.24.3
scipy==1.10.1
deap==1.3.3
streamlit==1.28.0
numpy==1.24.3
scipy==1.10.1
deap==1.3.3
google-api-python-client==2.84.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0